+++ changelog.md
@@ -3,3 +3,6 @@
 ## Changes
 - Blender recipe
 - Improve rocket rotation on launch pad
+
+## Fixes
+- Fix every gravity manipulator having the same animation